Add video controls to VideoPlayerController
The VideoPlayerController now includes media controls like "Play", "Pause", "Stop", and a time slider. It also allows the user to select a local video file to play. The VideoPlayerApp file was renamed to Main.java and the video-player.fxml file was substantially updated to reflect these changes in the application's user interface.